+++ assets/javascripts/kitten/components/cards/crowdfunding-card/components/informations.js
@@ -1,56 +1,6 @@
 import React, { PureComponent } from 'react'
-import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import { Text } from '../../../../components/typography/text'
-import { pxToRem } from '../../../../helpers/utils/typography'
-import COLORS from '../../../../constants/colors-config'
-import { ScreenConfig } from '../../../../constants/screen-config'
-
-const COMPONENT_GUTTER = pxToRem(10)
-
-const StyledContainer = styled.div`
-  margin-top: ${pxToRem(15)};
-  display: flex;
-  flex-wrap: wrap;
-  padding: 0;
-  line-height: 1;
-
-  @media (min-width: ${pxToRem(ScreenConfig.M.min)}) {
-    margin-top: ${pxToRem(20)};
-    padding: 0 ${COMPONENT_GUTTER};
-  }
-`
-
-const StyledInfo = styled.div`
-  margin-right: ${pxToRem(15)};
-  margin-bottom: ${pxToRem(5)};
-
-  @media (min-width: ${pxToRem(ScreenConfig.M.min)}) {
-    margin-right: ${pxToRem(20)};
-  }
-
-  :last-child {
-    margin-right: 0;
-  }
-`
-
-const StyledInfoLoading = styled.span`
-  display: block;
-  background-color: ${COLORS.line2};
-  border-bottom: ${pxToRem(1)} solid ${COLORS.background1};
-  width: ${pxToRem(40)};
-  height: ${pxToRem(16)};
-`
-
-const StyledInfoLoadingLarge = styled(StyledInfoLoading)`
-  width: ${pxToRem(65)};
-  border-top: ${pxToRem(1)} solid ${COLORS.background1};
-  border-bottom: 0;
-`
-
-const StyledText = styled(Text)`
-  font-size: ${pxToRem(12)};
-`
 
 class Informations extends PureComponent {
   static propTypes = {
@@ -73,11 +23,11 @@ class Informations extends PureComponent {
     if (!info1 && !info2 && !info3) return null
 
     return (
-      <StyledContainer>
+      <div className="k-CrowdfundingCard__informations">
         {this.renderInfo(info1)}
         {this.renderInfo(info2)}
         {this.renderInfo(info3)}
-      </StyledContainer>
+      </div>
     )
   }
 
@@ -87,20 +37,25 @@ class Informations extends PureComponent {
     if (!text) return null
 
     return (
-      <StyledInfo>
+      <div className="k-CrowdfundingCard__informations__infoContainer">
         {!loading && (
-          <StyledText lineHeight="normal" weight="light" color="font1">
+          <Text
+            lineHeight="normal"
+            weight="light"
+            color="font1"
+            className="k-CrowdfundingCard__informations__info"
+          >
             {text}
-          </StyledText>
+          </Text>
         )}
 
         {loading && (
           <>
-            <StyledInfoLoading />
-            <StyledInfoLoadingLarge />
+            <span className="k-CrowdfundingCard__informations__loadingElement" />
+            <span className="k-CrowdfundingCard__informations__loadingElement" />
           </>
         )}
-      </StyledInfo>
+      </div>
     )
   }
 }